Migrate to the new Entity layout
For details see https://github.com/bevyengine/bevy/pull/19121
https://github.com/bevyengine/bevy/pull/18704

The niche is now in the index, which makes the compression logic even
simpler.

The index now represented by NonMaxU32, which internally represented as
NonZeroU32 with all bits reversed, so we have to xor the bits.

I opened a PR to make it more ergonomic and avoid us relying on the
internal layout: https://github.com/bevyengine/bevy/pull/21246